fix: Exception 핸들러 수
